signal.c: fix type
* signal.c (sigsegv): type of write(2) is ssize_t.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@39447 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ signal.c
@@ -658,7 +658,7 @@ static RETSIGTYPE
 sigsegv(int sig SIGINFO_ARG)
 {
     if (segv_received) {
-	int RB_UNUSED_VAR(err);
+	ssize_t RB_UNUSED_VAR(err);
 	char msg[] = "SEGV received in SEGV handler\n";
 
 	err = write(2, msg, sizeof(msg));
